requirements.txt
openai
asyncio
python-dotenv
